correçao
concluido

+++ PROVA.py
@@ -1,7 +1,7 @@
-nota1=int(input('digite sua primeira nota:'))       
-nota2=int(input('digite sua segunda nota:'))
-nota3=int(input('digite sua terceira nota:'))
-nota4=int(input('digite sua quarta nota:'))
+nota1=float(input('digite sua primeira nota:'))       
+nota2=float(input('digite sua segunda nota:'))
+nota3=float(input('digite sua terceira nota:'))
+nota4=float(input('digite sua quarta nota:'))
 media= (nota1 + nota2 + nota3 + nota4) /4
 print ('sua média é',media)
